STD140Struct Lights Implementation

+++ res/shaders/screen.frag
@@ -1,7 +1,7 @@
 #version 450 core
 layout (std140, binding = 1) uniform WindowData
 {
-    vec2 windowSize;
+    ivec2 windowSize;
     float nearPlane;
     float farPlane;
     float gamma;
